+++ db/migrate/20210404152532_add_allow_weekends_to_randomization_schemes.rb
@@ -1,0 +1,5 @@
+class AddAllowWeekendsToRandomizationSchemes < ActiveRecord::Migration[6.0]
+  def change
+    add_column :randomization_schemes, :allow_tasks_on_weekends, :boolean, null: false, default: false
+  end
+end